requirements.txt
python<=3.8.0
tensorflow==2.3.0
keras==2.4.3
numpy>=1.21.5
opencv-python>=4.0.1
matplotlib>=3.5.1
segmentation_models==1.0.1
pillow>=9.0.1
seaborn==0.11.2